requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
email-validator==2.1.0
python-dotenv==1.0.0
WTForms==3.0.1
xhtml2pdf==0.2.15
gunicorn==21.2.0
requests==2.31.0
